refactor(format): extract summary rendering module
